Moved update commands to updates.py
